part_2/2_0_6/log-output-app/requirements.txt
flask==2.0.2
gunicorn==20.1.0
requests==2.26.0 
python-dotenv==0.19.2